requirements.txt
argcomplete==1.10.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
cattrs==1.8.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
compressed-rtf==1.0.6
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
executing==1.2.0
extract-msg==0.28.7
idna==3.4
IMAPClient==2.1.0
ipython==8.11.0
jedi==0.18.2
lxml==4.9.2
matplotlib-inline==0.1.6
olefile==0.46
parso==0.8.3
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycryptodome==3.17
Pygments==2.14.0
PyMuPDF==1.21.1
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
stack-data==0.6.2
traitlets==5.9.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wcwidth==0.2.6
xlrd==1.2.0
XlsxWriter==3.0.8
